Adding service code changes
This service is pretty straightforward, just a simple ‘get’ to the API. One thing to note is the property providedIn: 'root'. This is new to Angular 6 and it specifies what module the service is to be set up for, in this case, root. This allows for tree shaking to happen for services now.